requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.6
gunicorn==23.0.0
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.3.4
motor==3.7.1
dnspython==2.7.0
openai==1.51.2
azure-cognitiveservices-speech==1.40.0
numpy==1.26.4 
tenacity==8.5.0
pydub==0.25.1
pymupdf==1.24.9
typing-extensions>=4.7.0
azure-storage-blob==12.22.0
httpx==0.27.2
httpcore==1.0.5
bcrypt==4.3.0
ecdsa==0.19.1
passlib==1.7.4
pyasn1==0.6.1
python-jose==3.5.0
rsa==4.9.1